Split GumService into shared + platform partials (#3608)
Restructure the game-host MonoGameGum/GumService.cs monolith into a partial class spread across GumService.cs (shared ~90%), GumService.XnaLike.cs (#if XNALIKE), and GumService.Raylib.cs (#if RAYLIB), matching the codebase's cross-platform share-90%/if-the-divergence idiom. Pure behavior-preserving code motion: the compiled MonoGame/KNI/FNA/Raylib assemblies are unchanged.

Mid-method #if divergences become partial-void seams (AssignNativeTextInput, UninitializePlatform, ApplyTextureFilterPlatform, ExtractUnresolvedTextures); whole-body/platform-typed members (Initialize overloads, GameTime property + Update family, GetWindowSize, Draw(Camera2D), input factories) move into the platform partials. InitializeInternal splits into a per-platform bootstrap plus a shared FinishInitialize tail. The shared body carries no XNALIKE/RAYLIB #if — only a per-platform using block aliases the platform-namespaced types (Cursor/Sprite/CustomSetPropertyOnRenderable).

Remove the two dead #if !FRB guards: GumService.cs is absent from GumCoreShared.projitems, so it is never compiled with FRB defined and the guards were always-true dead code.

RaylibGum.csproj gains explicit Compile includes for the two partials; KniGum/FnaGum auto-glob them (the #if-wrapped inert partial is the safety belt). Adds headless XNALIKE seam pins (GumServicePartialSeamTests). Verified: MonoGame/KNI/Raylib build clean; MonoGameGum.Tests 1838/1838 and RaylibGum.Tests 467/467 green.

Follow-up to #3605; SilkNet fold-in tracked separately.
